requirements.txt
allure_python_commons==2.13.2
pytest==7.4.2
PyYAML==6.0.1
Requests==2.31.0
allure-pytest==2.13.2